Update V1.1.4
More optimization changes and a chnage to the Bounder interface format.

- interface.go:

Removed IsIntersect(), Min(), and Max() from Bounder as they only overcomplicated the interface structure.

- quadgo.go:

Moved IsIntersect to quadgo.go as isIntersect(). This know takes two entitys to check and returns a bool.

Code was also updated to use Bounds() instead of Min() and Max().

- bounds.go:

Removed IsIntersect(), Min(), and Max() as they were redundent and nolonger needed in Bounds.