package name fixed and changes to Objects
pixed package name to fallow conventoin. it is now quadgo instead of QuadGo.

chnages Insert to take a bounds and a variatic number of objects. i also changes Entity to hold a list of objects so that each Entity can have more then one object inside of it.

Changes Entity in the main QuadGo functions to be a pointer as its size is know posibly larger due to it holding a list of objects and i would be bad to pass around a copy of it all the time.