V3.0 final changes
Updated README to show the new changes to how the package works. Reworded a few comments to be more clear and updated some comments for Godoc's.

Changed the name of collapse to collapseAndMove and split to splitAndMove for clarity.

changes the code for zeroing out the entities and child nodes to use entities[:0] to falow conventions.

gofmt run.