Update 4.0.0 Final changes
Update 4.0.0

Overview:
This update makes a near complete overhaul of the backend of the library. Almost every single function has been changed in some way and a lot of the core types, Entity, Bound, and Point, have been changes in some way. Additional Concurrent read’s have been added. This means that Retrieve(), IsIntersect(), Intersects(), and IsEntity() are all run concurrently and return channels of their values rather than their values directly. One old bug has been addressed but also it has overall lowered performance of the library. It is now possible to have entities stored in more than one node at a time. This means that there is no longer issues with checking for intersects with large entities which could spread across multiple nodes. In the old version these entities were just deleted unknowingly and it was a problematic bug. In version 3.0.0 the library was changed to a center point based system, but this created a bug where a large entity would not be found for collision because its center point was put in only was node. Now on insert or move of an entity the entity bound is checked for collision with any of the 4 quadrants bounds. If it does collide its reference will be added to each node it collided with. This will address one problematic bug but also adds a large amount of extra overhead for basically all functions. Overall in small testes the overhead seems to have not created any big issues but larger scale benchmarks still need to be made to verify this. The plan is to add benchmarks and make any needed optimizations for V4.0.1 but currently the state of the code is worth pushing to a full version. There are also some other small changes to the library and some new support functions added which can be found in the godocs. To find more info on some of the new features check out the godocs and go back through the tutorial in the Readme to see how some of the new systems function.

Changed features:
All entity references are now added to all nodes it can fit in to, not just the node its center point fits in to.
New() now requires a given width and height as it was not good to have a default for something that should be required.
Max entities and max depth have been changed to uint as they should never be set to a negative number. This also makes the QuadGo struct take up a slightly smaller amount of space as instead of an int (which normally would be defaulted to int64 on 64 bit systems) it is now a uint16.
The library's “screen position” has been changed to having 0,0 at the top left of the screen instead of the bottom left of the screen. This is to follow the more conventional idea of 0,0 for computer screens.
All read actions have been updated to be run concurrently. This means that Retrieve(), IsIntersect(), Intersects(), and IsEntity() all return a channel type and are run on a separate thread, returning there values on that returned thread.

Added and changed features:
Added FindAndRemove() function to the Entities type.
Added Contains() to the Entities type.
Changed Entity to hold a func() Action instead of an empty interface.
Added NewEntityWithAction() for creating an entity with an action function
Added Entity.SetAction() for setting an action for an entity after creation.
Added Entity.IsEqual() function to check for equality of two entities.
Added Bound.IsEqual() function to check for equality of two bounds.
Added Point.IsEqual() function to check for equality of two points.

Check the commits from the V4.0 branch to see any other changes made to the library that may not be noted here.